Made express-notes-api modular
