+++ express-notes-api/data.json
@@ -1,5 +1,5 @@
 {
-  "nextId": 17,
+  "nextId": 19,
   "notes": {
     "2": {
       "id": 2,
@@ -34,12 +34,8 @@
       "id": 14
     },
     "15": {
-      "content": "I hope this is working.",
+      "content": "I hope this is working.??",
       "id": 15
-    },
-    "16": {
-      "content": "I hope this is working.",
-      "id": 16
     }
   }
 }
