+++ express-notes-api/write.js
@@ -7,4 +7,12 @@ function write(path, data) {
     }
   });
 }
-module.exports = write;
+
+function writeForDelete(path, data) {
+  fs.writeFile(path, JSON.stringify(data, null, 2), (err) => {
+    if (err) {
+      console.error('Error when writing', err);
+    }
+  });
+}
+module.exports = { write, writeForDelete };
